+++ apps/ui-dev/component/PageTools.jsx
@@ -1,11 +1,32 @@
-import { classnames } from "@elf-framework/sapa";
-import { Tools } from "@elf-framework/ui";
+import DarkModeFilled from "@elf-framework/icon/DarkModeFilled";
+import LightModeFilled from "@elf-framework/icon/LightModeFilled";
+import { classnames, useEffect, useState } from "@elf-framework/sapa";
+import {
+  Button,
+  IconButton,
+  RoundButton,
+  Switch,
+  Tools,
+} from "@elf-framework/ui";
 
 import mainMenus from "../constants/main-menus";
 import "./PageTools.scss";
 
 export function PageTools({ menu }) {
   const pathname = location.pathname;
+  const mode = localStorage.getItem("view-mode") || "light";
+  const [viewMode, setViewMode] = useState(mode);
+
+  useEffect(() => {
+    const mode = localStorage.getItem("view-mode") || "light";
+
+    if (mode === "light") {
+      document.body.classList.toggle("dark", false);
+    } else {
+      document.body.classList.toggle("dark", true);
+    }
+  }, []);
+
   return (
     <div class="page-tools">
       <div class="sm">
@@ -37,25 +58,55 @@ export function PageTools({ menu }) {
         />
       </div>
       <div class="lg">
-        {mainMenus.map((it, index) => {
-          if (index === 0) {
-            return <a href={it.link}>{it.title}</a>;
-          }
+        <div>
+          {mainMenus.map((it, index) => {
+            if (index === 0) {
+              return <a href={it.link}>{it.title}</a>;
+            }
 
-          const selected = pathname.startsWith(it.category);
+            const selected = pathname.startsWith(it.category);
 
-          return [
-            <span class="divider"></span>,
-            <a
-              href={it.link}
-              class={classnames({
-                selected,
-              })}
+            return [
+              <span class="divider"></span>,
+              <a
+                href={it.link}
+                class={classnames({
+                  selected,
+                })}
+              >
+                {it.title}
+              </a>,
+            ];
+          })}
+        </div>
+        <div style={{ justifyContent: "flex-end" }}>
+          {mode === "dark" ? (
+            <RoundButton
+              iconOnly
+              quiet
+              variant="dark"
+              onClick={() => {
+                setViewMode("light");
+                localStorage.setItem("view-mode", "light");
+                document.body.classList.toggle("dark", false);
+              }}
             >
-              {it.title}
-            </a>,
-          ];
-        })}
+              <DarkModeFilled />
+            </RoundButton>
+          ) : (
+            <RoundButton
+              iconOnly
+              quiet
+              onClick={() => {
+                setViewMode("dark");
+                localStorage.setItem("view-mode", "dark");
+                document.body.classList.toggle("dark", true);
+              }}
+            >
+              <LightModeFilled />
+            </RoundButton>
+          )}
+        </div>
       </div>
     </div>
   );