+++ apps/ui-dev/component/feedback/notification/MainView.jsx
@@ -15,7 +15,7 @@ export function MainView() {
     <VBox
       style={{
         gap: 30,
-        backgroundColor: "var(--color-gray-0)",
+        backgroundColor: "var(--color-gray-1)",
         padding: "60px 150px",
         position: "relative",
         height: 300,